+++ kwcoco/spec.py
@@ -1,6 +1,9 @@
 """
 TODO: need to investigate the correct way to use typing, what its limitations
 are, and what we have gained by carefully designing a type spec.
+
+TODO: We need to make this programatic so it can validate if a file is accepted
+or rejected. by the schema
 """
 
 import pathlib