+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 * Ported the non-torch components of CategoryTree from ndsampler to kwcoco. 
 * This helps kwcoco work better as a standalone package.
+* Union now works with videos
 
 
 ### Changed